+++ Conferentie/src/main/java/exceptions/DuplicateSprekerException.java
@@ -4,7 +4,7 @@ public class DuplicateSprekerException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
 
-	public DuplicateSprekerException(Integer id) {
+	public DuplicateSprekerException(Long id) {
 	    super("Duplicate speaker %s".formatted(id));
 	}
 }
