+++ Conferentie/src/main/java/exceptions/EvenementNotFoundException.java
@@ -4,7 +4,7 @@ public class EvenementNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
 
-	public EvenementNotFoundException(Integer id) {
+	public EvenementNotFoundException(Long id) {
 	    super("Could not find event %s".formatted(id));
 	  }
 }
