+++ Conferentie/src/main/java/exceptions/LokaalNotFoundException.java
@@ -4,7 +4,7 @@ public class LokaalNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
 
-	public LokaalNotFoundException(Integer id) {
+	public LokaalNotFoundException(Long id) {
 	    super("Could not find room %s".formatted(id));
 	  }
 }
