+++ Conferentie/src/main/java/exceptions/SprekerNotFoundException.java
@@ -4,7 +4,7 @@ public class SprekerNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
 
-	public SprekerNotFoundException(Integer id) {
+	public SprekerNotFoundException(Long id) {
 	    super("Could not find speaker %s".formatted(id));
 	  }
 }
